+++ packages/bloom/lib/commands/bloom/index.ts
@@ -1,4 +1,4 @@
-import type { RedisCommands, TypeMapping } from '@redis/client/dist/lib/RESP/types';
+import type { RedisCommands } from '@redis/client/dist/lib/RESP/types';
 
 import ADD from './ADD';
 import CARD from './CARD';
@@ -10,7 +10,8 @@ import MADD from './MADD';
 import MEXISTS from './MEXISTS';
 import RESERVE from './RESERVE';
 import SCANDUMP from './SCANDUMP';
-import { RESP_TYPES } from '@redis/client';
+
+export * from './helpers';
 
 export default {
   ADD,
@@ -34,31 +35,3 @@ export default {
   SCANDUMP,
   scanDump: SCANDUMP
 } as const satisfies RedisCommands;
-
-export function transformInfoV2Reply<T>(reply: Array<any>, typeMapping?: TypeMapping): T {
-  const mapType = typeMapping ? typeMapping[RESP_TYPES.MAP] : undefined;
-
-  switch (mapType) {
-    case Array: {
-      return reply as unknown as T;
-    }
-    case Map: {
-      const ret = new Map<string, any>();
-
-      for (let i = 0; i < reply.length; i += 2) {
-        ret.set(reply[i].toString(), reply[i + 1]);
-      }
-
-      return ret as unknown as T;
-    }
-    default: {
-      const ret = Object.create(null);
-
-      for (let i = 0; i < reply.length; i += 2) {
-        ret[reply[i].toString()] = reply[i + 1];
-      }
-
-      return ret as unknown as T;
-    }
-  }
-}
